requirements.txt
pandas
matplotlib
streamlit
protobuf==3.20.0
ipykernel

deltalake==0.19.2
openai==1.44.0
python-dotenv==1.0.1